First set of changes to switch from SimpleSite to TrimSite
SimpleSite has loads of competition on the web.  TrimSite doesn't, so
I'm in the process of renaming (pain, should have checked this before).
Also, have altered the line spacing in an ordered list so it looks
better - but have only done this where I currently need it, in
back-pic.css